core/levels/defender/audit/resources/register_func/requirements.txt
flask
sqlalchemy
google-api-python-client
google-auth
pg8000